# backend/src/json/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"version": "0.1",
		"title": "My_Little_Account_Book Backend server",
		"description": "나만의 작은 가계부 테스트 서버입니다. JWT기반의 인증방식을 가지고 있으므로 참고바랍니다."
	},
	"host": "localhost:3300",
	"basePath": "/",
	"schemes": [
		"http"
	],
	"tags": [
		{
			"name": "Auth",
			"description": "로그인, 회원가입 및 인증관련 API"
		}
	],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"description": "Access Token",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"cookieAuth": {
				"type": "apiKey",
				"description": "Refresh Token",
				"name": "refresh",
				"in": "header"
			}
		}
	},
	"paths": {
		"/auth/email": {
			"post": {
				"tags": [
					"Auth"
				],
				"description": "이메일 로그인",
				"requestBody": {
					"description": "이메일과 패스워드로 로그인",
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string",
										"example": "test@naver.com"
									},
									"password": {
										"type": "string",
										"example": "test123!@TEST"
									}
								},
								"required": [
									"email",
									"password"
								]
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "성공",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"refreshToken": {
											"type": "string",
											"description": "Refresh Token으로 Front에서 Cookie에 refresh를 키 값으로 설정 필요"
										},
										"accessToken": {
											"type": "string",
											"description": "Access Token으로 Front에서 Header의 Authorization의 값으로 설정 필요"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "아이디가 없거나 비밀번호가 틀림."
					},
					"403": {
						"description": "접근 권한 없음"
					}
				}
			}
		},
		"/auth/refresh": {
			"get": {
				"tags": [
					"Auth"
				],
				"description": "토큰 리프레쉬",
				"security": [
					{
						"cookieAuth": []
					},
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"description": "JWT의 Access token 갱신"
				},
				"responses": {
					"200": {
						"description": "성공",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"accessToken": {
											"type": "string",
											"description": "Access Token으로 Front에서 Header의 Authorization의 값으로 설정 필요"
										}
									}
								}
							}
						}
					},
					"403": {
						"description": "재 로그인 필요. 토큰 값이 만료됨."
					}
				}
			}
		}
	}
}